# openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Job to Invoice API",
    "version": "0.1.0",
    "description": "Owner authentication, workspace setup, and workspace-isolated draft jobs."
  },
  "paths": {
...
    "/v1/me": {
      "get": {
        "summary": "Owner bootstrap",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Verified identity, workspace version, and setup state" },
          "401": { "description": "Missing or invalid access token" },
          "503": { "description": "Database, role, or provisioning unavailable" }
        }
      }
    },
    "/v1/analytics/batch": {
      "post": {
        "summary": "Client analytics batch",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["events"],
                "properties": {
                  "events": { "type": "array", "minItems": 1, "maxItems": 50 }
                }
              }
            }
          }
        },
        "responses": {
          "202": { "description": "Accepted" },
          "401": { "description": "Missing or invalid access token" },
          "403": { "description": "Suspended or locked account" },
          "422": { "description": "Validation failed" }
        }
      }
    },
...
    "/webhooks/email": {
      "post": {
        "summary": "Verified Resend/Svix email delivery events",
        "security": [],
        "parameters": [
          {
            "name": "svix-id",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "svix-timestamp",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "svix-signature",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Verified event accepted or duplicate ignored" },
          "401": { "description": "Missing or invalid signature" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}